requirements.txt
streamlit
pillow
piexif
folium
streamlit-folium
